requirements.txt
asgiref @ file:///C:/b/abs_ae1n0p_8_w/croot/asgiref_1724072476516/work
chardet==5.2.0
comtypes==1.4.8
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
easyocr==1.7.2
et-xmlfile==1.1.0
filelock==3.16.1
fsspec==2024.10.0
gunicorn==23.0.0
imageio==2.36.0
inflection==0.5.1
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
pyclipper==1.3.0.post6
PyJWT==2.9.0
pypiwin32==223
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pyttsx3==2.98
pytz==2025.2
pywin32==308
PyYAML==6.0.2
reportlab==4.2.5
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.1.0
shapely==2.0.6
six==1.16.0
sqlparse==0.5.3
sympy==1.13.1
tifffile==2024.9.20
torch==2.5.0
torchvision==0.20.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
wheel==0.44.0
whitenoise==6.9.0
